+++ pylinac/plan_generator/dicom.py
@@ -15,7 +15,7 @@ from pydicom.uid import generate_uid
 
 from ..core.scale import wrap360
 from .fluence import plot_fluences
-from .mlc import STACK, DualStackMLCShaper, MLCShaper
+from .mlc import MLCShaper
 
 
 class GantryDirection(Enum):
@@ -392,14 +392,16 @@ class Beam:
         return self.ds
 
 
-class DualStackBeam:
-    """Represents a DICOM beam dataset. Has methods for creating the dataset and adding control points.
-    Generally not created on its own but rather under the hood as part of a PlanGenerator object.
-
-    It contains enough independent logic steps that it's worth separating out from the PlanGenerator class.
-    """
+class HalcyonBeam(Beam):
+    """A beam representing a Halcyon. Halcyons have dual MLC stacks, no X-jaws, no couch rotation, etc."""
 
     ds: Dataset
+    distal_mlc_boundaries = list(
+        range(-140, 141, 10)
+    )  # 141 because range end is exclusive
+    proximal_mlc_boundaries = list(
+        range(-145, 146, 10)
+    )  # 146 because range end is exclusive
 
     def __init__(
         self,
@@ -411,15 +413,13 @@ class DualStackBeam:
         machine_name: str,
         gantry_angles: float | list[float],
         gantry_direction: GantryDirection,
+        coll_angle: float,
         couch_vrt: float,
         couch_lat: float,
         couch_lng: float,
-        distal_mlc_boundaries: list[float],
-        proximal_mlc_boundaries: list[float],
         distal_mlc_positions: list[list[float]],
         proximal_mlc_positions: list[list[float]],
         metersets: list[float],
-        fluence_mode: FluenceMode,
     ):
         """
         Parameters
@@ -446,23 +446,25 @@ class DualStackBeam:
             The couch lateral position.
         couch_lng : float
             The couch longitudinal position.
+        distal_mlc_positions : list[list[float]]
+            The distal MLC positions for each control point. This is the x-position of each leaf for each control point.
+        proximal_mlc_positions : list[list[float]]
+            The proximal MLC positions for each control point. This is the x-position of each leaf for each control point.
         metersets : list[float]
             The meter sets for each control point. The length must match the number of control points in mlc_positions.
-        fluence_mode : FluenceMode
-            The fluence mode of the beam.
         """
         if len(beam_name) > 16:
             raise ValueError("Beam name must be less than or equal to 16 characters")
 
-        if beam_type == BeamType.STATIC and len(mlc_positions) != 1:
-            raise ValueError(
-                f"Static beam can only have one MLC position change ({len(mlc_positions)})"
-            )
+        if beam_type == BeamType.STATIC and (
+            len(proximal_mlc_positions) != 1 or len(distal_mlc_positions) != 1
+        ):
+            raise ValueError("Static beam can only have one MLC position change.")
 
-        if len(metersets) != len(mlc_positions):
+        if len(metersets) != len(proximal_mlc_positions) != len(distal_mlc_positions):
             raise ValueError(
                 f"The number of meter sets ({len(metersets)}) "
-                f"must match the number of MLC position changes ({len(mlc_positions)})"
+                f"must match the number of MLC position changes. Proximal: {len(proximal_mlc_positions)}; Distal: {len(distal_mlc_positions)}"
             )
 
         if (
@@ -479,10 +481,7 @@ class DualStackBeam:
         self.ds = self._create_basic_beam_info(
             beam_name,
             beam_type,
-            fluence_mode,
             machine_name,
-            num_leaves=len(mlc_positions[0]),
-            mlc_boundaries=mlc_boundaries,
         )
         if not isinstance(
             gantry_angles, Iterable
@@ -491,25 +490,25 @@ class DualStackBeam:
         self._append_initial_control_point(
             energy,
             dose_rate,
-            x1,
-            x2,
-            y1,
-            y2,
             gantry_angles[0],
             gantry_direction,
             coll_angle,
             couch_vrt,
             couch_lat,
             couch_lng,
-            couch_rot,
-            mlc_positions=mlc_positions[0],
+            proximal_mlc_positions=proximal_mlc_positions[0],
+            distal_mlc_positions=distal_mlc_positions[0],
         )
-        for mlc_pos, meterset, gantry_angle in zip(
-            mlc_positions[1:], metersets[1:], gantry_angles[1:]
+        for prox_mlc_pos, distal_mlc_pos, meterset, gantry_angle in zip(
+            proximal_mlc_positions[1:],
+            distal_mlc_positions[1:],
+            metersets[1:],
+            gantry_angles[1:],
         ):
             if beam_type == BeamType.DYNAMIC:
                 self._append_secondary_dynamic_control_point(
-                    mlc_positions=mlc_pos,
+                    proximal_mlc_positions=prox_mlc_pos,
+                    distal_mlc_positions=distal_mlc_pos,
                     meterset=meterset,
                     gantry_angle=gantry_angle,
                     gantry_dir=gantry_direction,
@@ -521,11 +520,9 @@ class DualStackBeam:
         self,
         beam_name: str,
         beam_type: BeamType,
-        fluence_mode: FluenceMode,
         machine_name: str,
-        num_leaves: int,
-        mlc_boundaries: list[float],
     ) -> Dataset:
+        """Halcyon-specific beam creation. We know the MLC boundaries, number of leaves, and that it's always FFF."""
         beam = Dataset()
         beam.Manufacturer = "Radformation"
         beam.ManufacturerModelName = "RadMachine"
@@ -538,14 +535,8 @@ class DualStackBeam:
         beam.PrimaryFluenceModeSequence = primary_fluence_mode_sequence
 
         primary_fluence_mode1 = Dataset()
-        if fluence_mode == FluenceMode.STANDARD:
-            primary_fluence_mode1.FluenceMode = "STANDARD"
-        elif fluence_mode == FluenceMode.FFF:
-            primary_fluence_mode1.FluenceMode = "NON_STANDARD"
-            primary_fluence_mode1.FluenceModeID = "FFF"
-        elif fluence_mode == FluenceMode.SRS:
-            primary_fluence_mode1.FluenceMode = "NON_STANDARD"
-            primary_fluence_mode1.FluenceModeID = "SRS"
+        primary_fluence_mode1.FluenceMode = "NON_STANDARD"
+        primary_fluence_mode1.FluenceModeID = "FFF"
 
         primary_fluence_mode_sequence.append(primary_fluence_mode1)
 
@@ -555,25 +546,33 @@ class DualStackBeam:
 
         # Beam Limiting Device Sequence: Beam Limiting Device 1
         beam_limiting_device1 = Dataset()
-        beam_limiting_device1.RTBeamLimitingDeviceType = "ASYMX"
+        beam_limiting_device1.RTBeamLimitingDeviceType = "X"
         beam_limiting_device1.NumberOfLeafJawPairs = "1"
         beam_limiting_device_sequence.append(beam_limiting_device1)
 
         # Beam Limiting Device Sequence: Beam Limiting Device 2
         beam_limiting_device2 = Dataset()
-        # TODO: likely that Elekta will need tweaking for this
-        beam_limiting_device2.RTBeamLimitingDeviceType = "ASYMY"
+        beam_limiting_device2.RTBeamLimitingDeviceType = "Y"
         beam_limiting_device2.NumberOfLeafJawPairs = "1"
         beam_limiting_device_sequence.append(beam_limiting_device2)
 
+        # Distal MLC
         # Beam Limiting Device Sequence: Beam Limiting Device 3
         beam_limiting_device3 = Dataset()
 
-        beam_limiting_device3.RTBeamLimitingDeviceType = "MLCX"
-        beam_limiting_device3.NumberOfLeafJawPairs = int(num_leaves / 2)
-        beam_limiting_device3.LeafPositionBoundaries = mlc_boundaries
-
+        beam_limiting_device3.RTBeamLimitingDeviceType = "MLCX1"
+        beam_limiting_device3.NumberOfLeafJawPairs = 28
+        beam_limiting_device3.LeafPositionBoundaries = self.distal_mlc_boundaries
         beam_limiting_device_sequence.append(beam_limiting_device3)
+
+        # Proximal MLC
+        # Beam Limiting Device Sequence: Beam Limiting Device 4
+        beam_limiting_device4 = Dataset()
+
+        beam_limiting_device4.RTBeamLimitingDeviceType = "MLCX2"
+        beam_limiting_device4.NumberOfLeafJawPairs = 29
+        beam_limiting_device4.LeafPositionBoundaries = self.proximal_mlc_boundaries
+        beam_limiting_device_sequence.append(beam_limiting_device4)
 
         # beam numbers start at 0 and increment from there.
         beam.BeamNumber = len(self.plan_ds.BeamSequence)
@@ -595,20 +594,16 @@ class DualStackBeam:
 
     def _append_initial_control_point(
         self,
-        energy: int,
+        energy: float,
         dose_rate: int,
-        x1: float,
-        x2: float,
-        y1: float,
-        y2: float,
         gantry_angle: float | list[float],
         gantry_rot: GantryDirection,
         coll_angle: float,
         couch_vrt: float,
         couch_lat: float,
         couch_lng: float,
-        couch_rot: float,
-        mlc_positions: list[float],
+        proximal_mlc_positions: list[float],
+        distal_mlc_positions: list[float],
     ):
         # Control Point Sequence: Control Point 0
         cp0 = Dataset()
@@ -622,29 +617,39 @@ class DualStackBeam:
 
         # Beam Limiting Device Position Sequence: Beam Limiting Device Position 1
         beam_limiting_device_position1 = Dataset()
-        beam_limiting_device_position1.RTBeamLimitingDeviceType = "ASYMX"
-        beam_limiting_device_position1.LeafJawPositions = [x1, x2]
+        beam_limiting_device_position1.RTBeamLimitingDeviceType = "X"
+        beam_limiting_device_position1.LeafJawPositions = [-140, 140]
         beam_limiting_device_position_sequence.append(beam_limiting_device_position1)
 
         # Beam Limiting Device Position Sequence: Beam Limiting Device Position 2
         beam_limiting_device_position2 = Dataset()
-        beam_limiting_device_position2.RTBeamLimitingDeviceType = "ASYMY"
-        beam_limiting_device_position2.LeafJawPositions = [y1, y2]
+        beam_limiting_device_position2.RTBeamLimitingDeviceType = "Y"
+        beam_limiting_device_position2.LeafJawPositions = [-140, 140]
         beam_limiting_device_position_sequence.append(beam_limiting_device_position2)
 
+        # Distal MLC
         # Beam Limiting Device Position Sequence: Beam Limiting Device Position 3
         beam_limiting_device_position3 = Dataset()
-        beam_limiting_device_position3.RTBeamLimitingDeviceType = "MLCX"
+        beam_limiting_device_position3.RTBeamLimitingDeviceType = "MLCX1"
         beam_limiting_device_position3.LeafJawPositions = [
-            f"{m:6f}" for m in mlc_positions
+            f"{m:6f}" for m in distal_mlc_positions
         ]  # convert to truncated string to fit VR limitations
         beam_limiting_device_position_sequence.append(beam_limiting_device_position3)
+
+        # Proximal MLC
+        # Beam Limiting Device Position Sequence: Beam Limiting Device Position 4
+        beam_limiting_device_position4 = Dataset()
+        beam_limiting_device_position4.RTBeamLimitingDeviceType = "MLCX2"
+        beam_limiting_device_position4.LeafJawPositions = [
+            f"{m:6f}" for m in proximal_mlc_positions
+        ]  # convert to truncated string to fit VR limitations
+        beam_limiting_device_position_sequence.append(beam_limiting_device_position4)
 
         cp0.GantryAngle = str(gantry_angle)
         cp0.GantryRotationDirection = gantry_rot.value
         cp0.BeamLimitingDeviceAngle = str(coll_angle)
         cp0.BeamLimitingDeviceRotationDirection = "NONE"
-        cp0.PatientSupportAngle = str(couch_rot)
+        cp0.PatientSupportAngle = "0"
         cp0.PatientSupportRotationDirection = "NONE"
         cp0.TableTopEccentricAngle = "0.0"
         cp0.TableTopEccentricRotationDirection = "NONE"
@@ -684,7 +689,8 @@ class DualStackBeam:
 
     def _append_secondary_dynamic_control_point(
         self,
-        mlc_positions: list[float],
+        proximal_mlc_positions: list[float],
+        distal_mlc_positions: list[float],
         meterset: float,
         gantry_angle: float,
         gantry_dir: GantryDirection,
@@ -704,13 +710,23 @@ class DualStackBeam:
         beam_limiting_device_position_sequence = Sequence()
         cp1.BeamLimitingDevicePositionSequence = beam_limiting_device_position_sequence
 
+        # Distal MLC
         # Beam Limiting Device Position Sequence: Beam Limiting Device Position 1
         beam_limiting_device_position1 = Dataset()
-        beam_limiting_device_position1.RTBeamLimitingDeviceType = "MLCX"
+        beam_limiting_device_position1.RTBeamLimitingDeviceType = "MLCX1"
         beam_limiting_device_position1.LeafJawPositions = [
-            f"{m:6f}" for m in mlc_positions
+            f"{m:6f}" for m in distal_mlc_positions
         ]  # convert to truncated string to fit VR limitations
         beam_limiting_device_position_sequence.append(beam_limiting_device_position1)
+
+        # Proximal MLC
+        # Beam Limiting Device Position Sequence: Beam Limiting Device Position 2
+        beam_limiting_device_position2 = Dataset()
+        beam_limiting_device_position2.RTBeamLimitingDeviceType = "MLCX2"
+        beam_limiting_device_position2.LeafJawPositions = [
+            f"{m:6f}" for m in proximal_mlc_positions
+        ]  # convert to truncated string to fit VR limitations
+        beam_limiting_device_position_sequence.append(beam_limiting_device_position2)
 
         # Referenced Dose Reference Sequence
         refd_dose_ref_sequence = Sequence()
@@ -718,10 +734,6 @@ class DualStackBeam:
 
         self.ds.NumberOfControlPoints = int(self.ds.NumberOfControlPoints) + 1
         self.ds.ControlPointSequence.append(cp1)
-
-    def as_dicom(self) -> Dataset:
-        """Return the beam as a DICOM dataset that represents a BeamSequence item."""
-        return self.ds
 
 
 class PlanGenerator:
@@ -949,7 +961,7 @@ class PlanGenerator:
         mlc = self._create_mlc()
         # create initial starting point; start under the jaws
         mlc.add_strip(
-            position_mm=strip_positions_mm[0] + 2,
+            position_mm=strip_positions_mm[0] - 2,
             strip_width_mm=strip_width_mm,
             meterset_at_target=0,
         )
@@ -1773,7 +1785,13 @@ class PlanGenerator:
         )
 
 
-class DualStackPlanGenerator(PlanGenerator):
+class STACK:
+    DISTAL = "distal"
+    PROXIMAL = "proximal"
+    BOTH = "both"
+
+
+class HalcyonPlanGenerator(PlanGenerator):
     """A class to generate a plan with two beams stacked on top of each other such as the Halcyon. This
     also assumes no jaws."""
 
@@ -1812,17 +1830,26 @@ class DualStackPlanGenerator(PlanGenerator):
         return self._old_beam.BeamLimitingDeviceSequence[-2].LeafPositionBoundaries
 
     @property
-    def proximal_leaf_config(self):
+    def proximal_leaf_config(self) -> list[float]:
         return self._old_beam.BeamLimitingDeviceSequence[-1].LeafPositionBoundaries
 
-    def _create_mlc(self) -> DualStackMLCShaper:
-        return DualStackMLCShaper(
-            distal_leaf_y_positions=self.distal_leaf_config,
-            proximal_leaf_y_positions=self.proximal_leaf_config,
+    def _create_mlc(self) -> (MLCShaper, MLCShaper):
+        """Create 2 MLC shaper objects, one for each stack."""
+        proximal_mlc = MLCShaper(
+            leaf_y_positions=self.proximal_leaf_config,
             max_x_mm=self.x_width / 2,
             sacrifice_gap_mm=self.sacrificial_gap,
-            max_overtravel_mm=self.max_sacrificial_move,
+            sacrifice_max_move_mm=self.max_sacrificial_move,
+            max_overtravel_mm=self.max_overtravel_mm,
         )
+        distal_mlc = MLCShaper(
+            leaf_y_positions=self.distal_leaf_config,
+            max_x_mm=self.x_width / 2,
+            sacrifice_gap_mm=self.sacrificial_gap,
+            sacrifice_max_move_mm=self.max_sacrificial_move,
+            max_overtravel_mm=self.max_overtravel_mm,
+        )
+        return proximal_mlc, distal_mlc
 
     def add_picketfence_beam(
         self,
@@ -1832,6 +1859,7 @@ class DualStackPlanGenerator(PlanGenerator):
         dose_rate: int = 600,
         energy: float = 6,
         gantry_angle: float = 0,
+        coll_angle: float = 0,
         couch_vrt: float = 0,
         couch_lng: float = 1000,
         couch_lat: float = 0,
@@ -1854,6 +1882,8 @@ class DualStackPlanGenerator(PlanGenerator):
             The energy of the beam.
         gantry_angle : float
             The gantry angle of the beam.
+        coll_angle : float
+            The collimator angle of the beam.
         couch_vrt : float
             The couch vertical position.
         couch_lng : float
@@ -1865,25 +1895,85 @@ class DualStackPlanGenerator(PlanGenerator):
         beam_name : str
             The name of the beam.
         """
-        # check MLC overtravel; machine may prevent delivery if exposing leaf tail
-        # max_dist_to_jaw = max(
-        #     max(abs(pos - x1), abs(pos + x2)) for pos in strip_positions_mm
-        # )
-        # if max_dist_to_jaw > self.max_overtravel_mm:
-        #     raise ValueError(
-        #         "Picket fence beam exceeds MLC overtravel limits. Lower padding, the number of pickets, or the picket spacing."
-        #     )
-        mlc = self._create_mlc()
-        # create initial starting point; start under the jaws
-        mlc.add_strip(
-            position_mm=strip_positions_mm[0] + 2,
-            strip_width_mm=strip_width_mm,
-            meterset_at_target=0,
+        prox_mlc, dist_mlc = self._create_mlc()
+
+        # we prepend the positions with an initial starting position 2mm from the first strip
+        # that way, each picket is the same cadence where the leaves move into position dynamically.
+        # If you didn't do this, the first picket might be different as it has the advantage
+        # of starting from a static position vs the rest of the pickets being dynamic.
+        strip_positions = [strip_positions_mm[0] - 2, *strip_positions_mm]
+        metersets = [0, *[1 / len(strip_positions_mm) for _ in strip_positions_mm]]
+
+        for strip, meterset in zip(strip_positions, metersets):
+            if stack in (STACK.DISTAL, STACK.BOTH):
+                dist_mlc.add_strip(
+                    position_mm=strip,
+                    strip_width_mm=strip_width_mm,
+                    meterset_at_target=meterset,
+                )
+                if stack == STACK.DISTAL:
+                    prox_mlc.park(meterset=meterset)
+            if stack in (STACK.PROXIMAL, STACK.BOTH):
+                prox_mlc.add_strip(
+                    position_mm=strip,
+                    strip_width_mm=strip_width_mm,
+                    meterset_at_target=meterset,
+                )
+                if stack == STACK.PROXIMAL:
+                    dist_mlc.park(meterset=meterset)
+
+        beam = HalcyonBeam(
+            plan_dataset=self.ds,
+            beam_name=beam_name,
+            beam_type=BeamType.DYNAMIC,
+            energy=energy,
+            dose_rate=dose_rate,
+            gantry_angles=gantry_angle,
+            gantry_direction=GantryDirection.NONE,
+            coll_angle=coll_angle,
+            couch_vrt=couch_vrt,
+            couch_lat=couch_lat,
+            couch_lng=couch_lng,
+            proximal_mlc_positions=prox_mlc.as_control_points(),
+            distal_mlc_positions=dist_mlc.as_control_points(),
+            metersets=prox_mlc.as_metersets(),  # can use either MLC for metersets
+            machine_name=self.machine_name,
+        )
+        self.add_beam(beam.as_dicom(), mu=mu)
+
+    def add_open_field_beam(
+        self,
+    ):
+        """Add an open field beam to the plan. This can be defined by one or both MLC stacks.
+        The x1, x2, y1, and y2 terms are colloquial. The MLCs define the field edges.
+        The nearest MLC to the field edge in the y-direction will be the one used. It can round up or down.
+        """
+        raise NotImplementedError(
+            "Open field beams are not yet implemented for Halcyon plans"
         )
 
-        # for strip in strip_positions_mm:
-        #     # starting control point
-        #     mlc.add_strip(
-        #         position_mm=strip,
-        #         strip_width_mm=strip_width_mm,
-        #         meterset_at
+    def add_dose_rate_beams(
+        self,
+    ):
+        raise NotImplementedError(
+            "Dose rate beams are not yet implemented for Halcyon plans"
+        )
+
+    def add_mlc_speed_beams(self):
+        raise NotImplementedError(
+            "MLC speed beams are not yet implemented for Halcyon plans"
+        )
+
+    def add_gantry_speed_beams(
+        self,
+    ):
+        raise NotImplementedError(
+            "Gantry speed beams are not yet implemented for Halcyon plans"
+        )
+
+    def add_winston_lutz_beams(
+        self,
+    ):
+        raise NotImplementedError(
+            "Winston-Lutz beams are not yet implemented for Halcyon plans"
+        )