+++ docs/source/changelog.rst
@@ -2,6 +2,16 @@
 Changelog
 =========
 
+v 3.25.0
+--------
+
+Plan Generator
+^^^^^^^^^^^^^^
+
+* The plan generator can now generate Halcyon picket fence plans using the new :class:`~pylinac.planar_imaging.HalcyonPlanGenerator` class.
+  Currently, this specialized generator is limited to picket fence plans, but more plan types will be added in the future. See :ref:`halcyon-plan-generator`.
+* The :class:`MLCShaper` has a new method: ``park``.
+
 v 3.23.2
 --------